feat(balancer) use dynamic upstream keepalive pools
This replaces the keepalive implementation from ngx_http_upstream_module
with the one proposed by
https://github.com/openresty/lua-nginx-module/pull/1600.

For the time being, the ngx_http_lua and lua-resty-core patches have
been applied to our openresty-build-tools repository and are applied
when building OpenResty for Kong. The final API may be subject to
changes by OpenResty, but incorporating the changes in Kong ahead of
time reduces possible virtual host confusion issues and/or attack
vectors.

The proposed OpenResty patches offer a way to circumvent the NGINX
limitations detailed in the following tickets:

* https://trac.nginx.org/nginx/ticket/1340
* https://trac.nginx.org/nginx/ticket/1593

The behavior changes in Kong from this work are:

* Avoid virtual host confusion when proxying upstream virtual services
  over TLS (as of today, only based on the SNI).
* Allow for an indefinite idle timeout and max requests for connections
  in the keepalive pool (a behavior reverted by NGINX in 1.15.3).

In the future, we could implement per-Service (or per-Upstream)
keepalive pools, allowing users to granularly enable/disable, or
configure upstream keepalive pools as they wish.

Fix #2835
Fix #3788
Fix #4001